refactor: useChangeServer 훅 분리 및 적용
